requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
diff-match-patch==20241021
Django==5.1.6
django-colorfield==0.12.0
django-cors-headers==4.7.0
django-import-export==4.3.7
django-jazzmin==3.0.1
django-recaptcha==4.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
pillow==11.1.0
pycountry==24.6.1
PyJWT==2.10.1
python-slugify==8.0.4
requests==2.32.3
sqlparse==0.5.3
tablib==3.8.0
text-unidecode==1.3
tzdata==2025.1
urllib3==2.3.0
